TSF Task 1- DS and BA
Prediction using Supervised ML
Predict the percentage of an student based on the no. of study hours.
This is a simple linear regression task as it involves just 2 variables.
You can use R, Python, SAS Enterprise Miner or any other tool 
What will be predicted score if a student studies for 9.25 hrs/ day?